aider: Add the appEnv to the testCases struct in `TestGetSuggestions`.
# Aider chat conversation:

USER: Add the appEnv to the testCases struct in `TestGetSuggestions`

USER: Add the appEnv to the testCases struct in `TestGetSuggestions`

ASSISTANT: To add the `appEnv` to the `testCases` struct in `TestGetSuggestions`, we need to modify the `testCases` struct to include an `appEnv` field. Then, we can use this field in the test loop.

Here is the *SEARCH/REPLACE block* for the changes:

```go
internal/db/sanitise_test.go
<<<<<<< SEARCH
	testCases := []struct {
		name           string
		input          []suggestionsResults
		expectedOutput []Suggestion
	}{
=======
	testCases := []struct {
		name           string
		appEnv         AppEnv
		input          []suggestionsResults
		expectedOutput []Suggestion
	}{
		{
			name:   "Sanitising test",
			appEnv: appEnv,
>>>>>>> REPLACE
```

```go
internal/db/sanitise_test.go
<<<<<<< SEARCH
	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			suggestions := getSuggestions(appEnv, tc.input, 0)
=======
	for _, tc := range testCases {
		t.Run(tc.name, func(t *testing.T) {
			suggestions := getSuggestions(tc.appEnv, tc.input, 0)
>>>>>>> REPLACE
```

This change will allow each test case to use a different `appEnv`, which can be useful for testing different configurations or states of the application environment.