aider: Replace incorrect `bookmarks` field with correct `episodeDetails` field in `builder_test.go`.
# Aider chat conversation:

ASSISTANT: The error messages indicate that there are struct literals in the `builder_test.go` file that are incorrectly using a field named `bookmarks`, which does not exist in the struct definition. Instead, it seems that the field should be `episodeDetails`, which is of type `internal.EpisodeDetails`.

To fix this, we need to replace the incorrect `bookmarks` field with the correct `episodeDetails` field in the test cases within the `builder_test.go` file.

Here are the changes that need to be made:

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			name: "Test Case 1",
			bookmarks: []internal.Bookmark{
=======
			name: "Test Case 1",
			episodeDetails: internal.EpisodeDetails{
>>>>>>> REPLACE
```

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			name: "Renaming Deadworld",
			bookmarks: []internal.Bookmark{
=======
			name: "Renaming Deadworld",
			episodeDetails: internal.EpisodeDetails{
>>>>>>> REPLACE
```

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			name: "Strontium Dog",
			bookmarks: []internal.Bookmark{
=======
			name: "Strontium Dog",
			episodeDetails: internal.EpisodeDetails{
>>>>>>> REPLACE
```

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			name: "Strontium Dug",
			bookmarks: []internal.Bookmark{
=======
			name: "Strontium Dug",
			episodeDetails: internal.EpisodeDetails{
>>>>>>> REPLACE
```

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			name: "ABC Warriors",
			bookmarks: []internal.Bookmark{
=======
			name: "ABC Warriors",
			episodeDetails: internal.EpisodeDetails{
>>>>>>> REPLACE
```

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			name: "The ABC Warriors",
			bookmarks: []internal.Bookmark{
=======
			name: "The ABC Warriors",
			episodeDetails: internal.EpisodeDetails{
>>>>>>> REPLACE
```

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			details := make([]internal.EpisodeDetails, len(tc.bookmarks))
			for i, b := range tc.bookmarks {
=======
			details := make([]internal.EpisodeDetails, len(tc.episodeDetails))
			for i, b := range tc.episodeDetails {
>>>>>>> REPLACE
```

```go
internal/scanner/builder_test.go
<<<<<<< SEARCH
			issue := buildIssue(appEnv, "2000AD 123 (1977).pdf", details)
			assert.Equal(t, 123, issue.IssueNumber)
			assert.Equal(t, tc.expectedSeries, issue.Episodes[0].Series.Title)
			assert.Equal(t, tc.expectedTitle, issue.Episodes[0].Title)
			assert.Equal(t, tc.expectedPart, issue.Episodes[0].Part)
=======
			issue := buildIssue(appEnv, "2000AD 123 (1977).pdf", tc.episodeDetails)
			assert.Equal(t, 123, issue.IssueNumber)
			assert.Equal(t, tc.expectedSeries, issue.Episodes[0].Series.Title)
			assert.Equal(t, tc.expectedTitle, issue.Episodes[0].Title)
			assert.Equal(t, tc.expectedPart, issue.Episodes[0].Part)
>>>>>>> REPLACE
```

Please note that the above changes assume that the `episodeDetails` field in the struct literal should contain a single `internal.EpisodeDetails` object. If the `episodeDetails` field is supposed to be a slice of `internal.EpisodeDetails` objects (similar to the original `bookmarks` slice), then you will need to adjust the `episodeDetails` field to be a slice and populate it accordingly. If that's the case, please let me know, and I can provide the updated changes.