Download only selected progs (#15)
* Download list of progs and display

This is the start of choosing which progs to download. Next step is to
cache that data so that I can stop hitting the Rebellion site with
dozens of requests for every test.

* Save downloaded prog list to disc

* Add scribble dependency

* Load saved progs on startup

* Move towards a better download list

* Show which progs are already downloaded

* Hide downloaded progs

* Show downloaded progs

* Download latest prog and update list

* Move api methods to api layer

We want Megazines in future, so I think delegating to the struct makes
sense. Still not entirely sure of go inheritance and composition.

* Download selected progs

* Option for only listing the latest